+++ android/src/main/kotlin/com/kiwi/fluttercrashlytics/FlutterCrashlyticsPlugin.kt
@@ -1,7 +1,9 @@
 package com.kiwi.fluttercrashlytics
 
-import android.app.Activity
+import android.content.Context
 import android.content.Intent
+import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
+import android.util.Log
 import com.crashlytics.android.Crashlytics
 import io.fabric.sdk.android.Fabric
 import io.flutter.plugin.common.MethodCall
@@ -10,13 +12,12 @@ import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry.Registrar
 
-
-class FlutterCrashlyticsPlugin(private val context: Activity) : MethodCallHandler {
+class FlutterCrashlyticsPlugin(private val context: Context) : MethodCallHandler {
     companion object {
         @JvmStatic
         fun registerWith(registrar: Registrar) {
             val channel = MethodChannel(registrar.messenger(), "flutter_crashlytics")
-            channel.setMethodCallHandler(FlutterCrashlyticsPlugin(registrar.activity()))
+            channel.setMethodCallHandler(FlutterCrashlyticsPlugin(registrar.context()))
         }
     }
 
@@ -46,7 +47,7 @@ class FlutterCrashlyticsPlugin(private val context: Activity) : MethodCallHandle
                     core.log(call.arguments as String)
                 } else {
                     val info = call.arguments as List<Any>
-                    core.log(info[0] as Int, info[1] as String, info[2] as String)
+                    core.log(info[0].toString() + ": " + info[1] + " " + info[2])
                 }
                 result.success(null)
             }
@@ -86,9 +87,10 @@ class FlutterCrashlyticsPlugin(private val context: Activity) : MethodCallHandle
                     //Start a new activity to not crash directly under onMethod call, or it will crash JNI instead of a clean exception
                     val intent = Intent(context, CrashActivity::class.java).apply {
                         putExtra("exception", throwable)
+                        flags = FLAG_ACTIVITY_NEW_TASK
                     }
 
-                    context.startActivityForResult(intent, -1)
+                    context.startActivity(intent)
                 } else {
                     core.logException(throwable)
                 }