+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-## 0.1.3
+## 0.2.1
 
-* Dependencies updated (crashlytics 2.9.7, gradle 3.2.1, kotlin 1.3.11)
+* Dependencies updated (crashlytics 2.9.8, gradle 3.3.0, kotlin 1.3.20)
+
+## 0.2.0
+* move from swift to obj-c
+
+## 0.1.3
+* fix android background execution
+
+## 0.1.2
+* fix iOS custom log
 
 ## 0.1.0
 * Breaking change, now initialize method need to be called in order to init Fabric (GDPR compliance)